+++ libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: maneddam <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: eej-jama <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 00:23:09 by eej-jama          #+#    #+#             */
-/*   Updated: 2023/03/25 10:35:14 by maneddam         ###   ########.fr       */
+/*   Updated: 2023/03/30 15:23:56 by eej-jama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,30 +17,30 @@
 # include <unistd.h>
 
 int					ft_atoi(const char *str);
-void				ft_bzero(void *s, size_t n);
-void				*ft_calloc(size_t count, size_t size);
+void				ft_bzero(void *s, int n);
+void				*ft_calloc(int count, int size);
 int					ft_isalnum(int c);
 int					ft_isalpha(int c);
 int					ft_isascii(int c);
 int					ft_isdigit(int c);
 int					ft_isprint(int c);
-void				*ft_memchr(const void *s, int c, size_t n);
-int					ft_memcmp(const void *s1, const void *s2, size_t n);
-void				*ft_memcpy(void *dst, const void *src, size_t n);
-void				*ft_memmove(void *dst, const void *src, size_t len);
-void				*ft_memset(void *b, int c, size_t len);
+void				*ft_memchr(const void *s, int c, int n);
+int					ft_memcmp(const void *s1, const void *s2, int n);
+void				*ft_memcpy(void *dst, const void *src, int n);
+void				*ft_memmove(void *dst, const void *src, int len);
+void				*ft_memset(void *b, int c, int len);
 char				*ft_strchr(const char *s, int c);
 char				*ft_strdup(const char *s1);
-size_t				ft_strlcat(char *dst, const char *src, size_t dstsize);
-size_t				ft_strlcpy(char *dst, const char *src, size_t dstsize);
-size_t				ft_strlen(const char *s);
-int					ft_strncmp(const char *s1, const char *s2, size_t n);
+int				ft_strlcat(char *dst, const char *src, int dstsize);
+int				ft_strlcpy(char *dst, const char *src, int dstsize);
+int				ft_strlen(const char *s);
+int					ft_strncmp(const char *s1, const char *s2, int n);
 char				*ft_strnstr(const char *haystack, const char *needle,
-						size_t len);
+						int len);
 char				*ft_strrchr(const char *s, int c);
 int					ft_tolower(int c);
 int					ft_toupper(int c);
-char				*ft_substr(char const *s, unsigned int start, size_t len);
+char				*ft_substr(char const *s,  int start, int len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s1, char const *set);
 char				**ft_split(char const *s, char c);
